# Optimize `Mesh::add_rect_with_uv`