<!--
Please read the "Making a PR" section of [`CONTRIBUTING.md`](https://github.com/emilk/egui/blob/main/CONTRIBUTING.md) before opening a Pull Request!

* Keep your PR:s small and focused.
* The PR title is what ends up in the changelog, so make it descriptive!
* If applicable, add a screenshot or gif.
* If it is a non-trivial addition, consider adding a demo for it to `egui_demo_lib`, or a new example.
* Do NOT open PR:s from your `master` branch, as that makes it hard for maintainers to test and add commits to your PR.
* Remember to run `cargo fmt` and `cargo clippy`.
* Open the PR as a draft until you have self-reviewed it and run `./scripts/check.sh`.
* When you have addressed a PR comment, mark it as resolved.

Please be patient! I will review your PR, but my time is limited!
-->

* [x] I have followed the instructions in the PR template


While optimizing #7431, I noticed that `Mesh::add_rect_with_uv` was taking quite a bit of time. Using `extend_with_slice` results in much better codegen (fewer branches) and inlining the function makes things faster still. The same applies to `Mesh::add_triangle`.

Current `main`:

```
text_layout_uncached    time:   [65.988 µs 66.042 µs 66.108 µs]
                        change: [+0.1351% +0.3381% +0.5863%] (p = 0.00 < 0.05)
```

With these optimizations:
```
text_layout_uncached    time:   [59.996 µs 60.104 µs 60.269 µs]
                        change: [−9.8641% −9.6530% −9.4583%] (p = 0.00 < 0.05)
                        Performance has improved.
```